requirements.txt
# Azure IoT Device SDK
azure-iot-device>=2.12.0

# HTTP client for updates
aiohttp>=3.8.0

# System monitoring (optional but recommended)
psutil>=5.9.0

# Jetson Nano specific monitoring (optional)
# jetson-stats>=4.2.0  # Uncomment if running on Jetson Nano

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For advanced logging
colorlog>=6.7.0 